Fix a nasty bug in CameraData serialization/deserialization.
The 'size_t' data type has a different size on my laptop (8 bits)
than it does on the Raspberry Pi 2 (4 bits). This caused the
the spacing of data in the zmq message to be different on my laptop
than on the Pi, leading to deserialization failure. The fix was
to switch all datatypes that get serialized to types of explicit size,
e.g. uint8_t and uint16_t.
Also added an extra form of data validation in these zmq messages -
each message type has its own "validation code", which provides
immediate verification that the received message is indeed a
serialized vector of CameraData.

+++ src/core/types/CameraSpecifications.h
@@ -2,6 +2,7 @@
 #define OBSTACLE_DETECTION_ICAMERASPECIFICATIONS_H
 
 #include <stddef.h>
+#include <cstdint>
 
 /**
  * A data type for storing the parameters of cameras.
@@ -16,9 +17,9 @@
  */
 struct CameraSpecifications {
 public:
-    const unsigned int pixelHeight;
-    const unsigned int pixelWidth;
-    const size_t bytesPerPixel;
+    const uint32_t pixelHeight;
+    const uint32_t pixelWidth;
+    const uint8_t bytesPerPixel;
 
     const double FOVDegreesVertical;
     const double FOVDegreesHorizontal;
